Changes to make two rec methods match, including sorting and additional debug outputs
