+++ papas/aliceproto/DAG.py
@@ -34,14 +34,14 @@ class Node(object):
     each node has an arbitrary number of children and parents
     There are no loops in the directed DAG
     But there may be loops in the undirected version of the DAG
-    
+
     attributes:
        value = the item of interest (around which the node is wrapped)
        children = list of child nodes
        parents  = list of parent node
        undirected_links = combined list of parents and children
     '''
-    
+
     def __init__(self, value):
         '''constructor. 
         value can be anything, even a complex object. 
@@ -52,13 +52,13 @@ class Node(object):
         self.children = []
         self.parents = []
         self.undirected_links = [] #the union of the parents and children (other implementations possible)
-        
+
     def get_value(self):
         return self.value
-    
+
     def accept(self, visitor):
         visitor.visit(self)
-             
+
     def add_child(self, child):
         '''set the children'''
         self.children.append(child)
@@ -69,7 +69,7 @@ class Node(object):
         '''set the parents'''
         self.parents.append(parent)
         self.undirected_links.append(parent)
-        
+
     def remove_all_links_to(self,toremove):
         '''checks for element toremove in the list of children and parents and
            removes any links from both this and from the toremove node
@@ -82,7 +82,7 @@ class Node(object):
             toremove.parents.remove(self)        
 
     def get_linked_nodes(self, type):  #ask colin, I imagine there is a more elegant Python way to do this
-                                       #alice todo make type a enumeration and not a string?
+                                        #alice todo make type a enumeration and not a string?
         '''return a list of the linked children/parents/undirected links'''
         if (type is "children"):
             return self.children
@@ -96,28 +96,28 @@ class Node(object):
         return self.__str__()
 
     def __str__(self):
-            '''unique string representation'''         
-            return str('node: {val} {children}'.format(
-                val = self.value,
-                children = self.children
-                ) )    
+        '''unique string representation'''         
+        return str('node: {val} {children}'.format(
+            val = self.value,
+            children = self.children
+        ) )    
 
 
 class BreadthFirstSearch(object):
-    
+
     def __init__(self,root, link_type):
         '''Perform the breadth first recursive search of the nodes'''
         self.result = []
         self.root = root
         self.visited = dict()
         self.bfs_recursive([root],link_type)
-        
+
     def visit(self, node):
         if self.visited.get(node, False):
             return
         self.result.append( node )
         self.visited[node] = True
-             
+
     def bfs_recursive(self,nodes, link_type ):
         '''Breadth first recursive implementation
         each recursion is one level down the tree
@@ -125,7 +125,7 @@ class BreadthFirstSearch(object):
         link_nodes = []
         if len(nodes) is 0:
             return 
-        
+
         for node in nodes: # collect a list of all the next level of nodes
             if (self.visited.get(node, False)):  
                 continue              
@@ -134,17 +134,12 @@ class BreadthFirstSearch(object):
             if (self.visited.get(node, False)):  
                 continue            
             node.accept(self)
-        
+
         self.bfs_recursive(link_nodes,  link_type)
-        
-     
 
-         
-     
 
-          
 class BreadthFirstSearchIterative(object):
-    
+
     def __init__(self,root, link_type):
         '''Perform the breadth first iterative search of the nodes'''
         self.visited = {}
@@ -156,16 +151,16 @@ class BreadthFirstSearchIterative(object):
             return
         self.result.append( node )
         self.visited[node] = True
-        
+
     def bfs_iterative(self,node, link_type ):
         '''Breadth first iterative implementation
         using a deque to order the nodes 
         link_type can be "children", "parents","undirected" '''
-            
+
         # Create a deque for the Breadth First Search
         todo = deque()
         todo.append( node)
-        
+
         while len(todo):
             node = todo.popleft()
             if self.visited.get(node,False): #check if already processed
@@ -175,10 +170,10 @@ class BreadthFirstSearchIterative(object):
                 if self.visited.get(linknode,False): #check if already processed
                     continue
                 todo.append( linknode)  
-                
-                
+
+
 class DAGFloodFill(object):
-    
+
     def __init__(self, elements, first_label = 1):
         '''Iterate through all nodes and 
         use Breadth first search to find connected groups'''
@@ -187,18 +182,18 @@ class DAGFloodFill(object):
         self.visited = dict()
         self.blocks = []
         for uid, node in elements.iteritems():
-                if self.visited.get(node, False): #already done so skip the rest
-                    continue
-                
-                #find connected nodes
-                bfs = BreadthFirstSearchIterative(node,"undirected")
-                
-                # set all connected elements to have a visited flag =true
-                for n in bfs.result :
-                    self.visited.update({n: True})
-                #add into the set of blocks
-                self.blocks.append( bfs.result)
-                self.label += 1
-        
-        
-              
+            if self.visited.get(node, False): #already done so skip the rest
+                continue
+
+            #find connected nodes
+            bfs = BreadthFirstSearchIterative(node,"undirected")
+
+            # set all connected elements to have a visited flag =true
+            for n in bfs.result :
+                self.visited.update({n: True})
+            #add into the set of blocks
+            self.blocks.append( bfs.result)
+            self.label += 1
+
+
+
