+++ papas/aliceproto/blocksplitter.py
@@ -1,7 +1,7 @@
 from blockbuilder import BlockBuilder
 
 class BlockSplitter(BlockBuilder):
-    ''' BlockSplitter takes a block of particle flow element ids (clusters,tracks etc)
+    ''' BlockSplitter takes an exisiting block of particle flow element ids (clusters,tracks etc)
         and unlinks any specified edges. It then recalculates a new set of sub-blocks
         using the underlying BlockBuilder class
         
@@ -25,15 +25,19 @@ class BlockSplitter(BlockBuilder):
         super(BlockSplitter, self).__init__(block.element_uniqueids, block.edges, history_nodes, block.pfevent)
         assert( isinstance(self.blocks,dict))
             
-        #the new blocks are subblocks of the original block
-        #so keep note of this in the history (at least for now)
-        # nb in some cases the new block will be the same as the original block although
-        # the edges will have changed (for python these changes will also be seen in
-        # the original block)        
-        if (self.history_nodes != None ):
-            for subblockid in self.blocks.keys():
-                #print "split" , block.uniqueid, subblockid
-                self.history_nodes[block.uniqueid].add_child(history_nodes[subblockid])  
+        #Plan 1 was to link new blocks to original blocks. But this makes use of the history nodes
+        # rather cumbersome. needs to be debated further removed for now
+        ##the new blocks are subblocks of the original block
+        ##so keep note of this in the history (at least for now)
+        ## nb in some cases the new block will be the same as the original block although
+        ## the edges will have changed (for python these changes will also be seen in
+        ## the original block)        
+        #if (self.history_nodes != None ):
+            #for subblockid in self.blocks.keys():
+                #if len(block.element_uniqueids)>5:
+                    #print "*****split", block
+                    #pass
+                ##self.history_nodes[block.uniqueid].add_child(history_nodes[subblockid])  
                  
         #set the original block to be inactive
         block.is_active = False
