+++ papas/pfalgo/links.py
@@ -110,7 +110,7 @@ class Links(dict):
         lines = []
         for key, val in self.iteritems():
             ele1, ele2 = key
-            lines.append("{ele1:50} {ele2:50} {val:5.4f}".format(ele1=ele1,
+            lines.append("{ele1:50} {ele2:50} dist = {val:5.4f}".format(ele1=ele1,
                                                                  ele2=ele2,
                                                                  val=val))
         '\n Groups:\n'.join(lines)
